core/urls.py
from django.urls import path, include
from rest_framework.routers import DefaultRouter

from core import views

router = DefaultRouter()
router.register(r'categories', views.CategoryViewSet, basename='category')
router.register(r'products', views.ProductViewSet, basename='product')
router.register(r'orders', views.OrderViewSet, basename='order')

urlpatterns = [
    path('', include(router.urls)),
    path('order-items/', views.OrderItemCreateView.as_view(), name='order-items'),
    path('revenue', views.RevenueAPIView.as_view(), name='revenue')
]
