requirements.txt
Django>=3.0,<4.0
psycopg2-binary>=2.8
djangorestframework
dj-rest-auth
django-allauth[with_social]
djangorestframework-simplejwt
Pillow
drf-yasg
django-money[exchange]
factory_boy
coverage